+++ lib/components/Information/exports.tsx
@@ -1,3 +1,1 @@
-export * from "./Alert"
-export * from "./Avatar"
 export * from "./Badge"
