+++ lib/experimental/Information/Avatar/index.stories.tsx
