+++ lib/experimental/Information/Avatar/index.tsx
