+++ lib/experimental/Information/exports.tsx
@@ -1,0 +1,4 @@
+export * from "./Alert"
+export * from "./Avatar"
+export * from "./Counter"
+export * from "./Shortcut"
